go back to 11485

+++ Source/UnitTests/Common/FloatUtilsTest.cpp
@@ -1,5 +1,6 @@
 // Copyright 2018 Dolphin Emulator Project
-// SPDX-License-Identifier: GPL-2.0-or-later
+// Licensed under GPLv2+
+// Refer to the license.txt file included.
 
 #include <limits>
 #include <random>
@@ -28,8 +29,10 @@ TEST(FloatUtils, FlushToZero)
   // we want the multiplication to occur at test runtime.
   volatile float s = std::numeric_limits<float>::denorm_min();
   volatile double d = std::numeric_limits<double>::denorm_min();
-  EXPECT_LT(0.f, s * 2);
-  EXPECT_LT(0.0, d * 2);
+  // Casting away the volatile attribute is required in order for msvc to resolve this to the
+  // correct instance of the comparison function.
+  EXPECT_LT(0.f, (float)(s * 2));
+  EXPECT_LT(0.0, (double)(d * 2));
 
   EXPECT_EQ(+0.0, Common::FlushToZero(+std::numeric_limits<double>::denorm_min()));
   EXPECT_EQ(-0.0, Common::FlushToZero(-std::numeric_limits<double>::denorm_min()));
